requirements.txt
Django==1.9.4
argparse==1.2.1
dj-database-url==0.4.0
django-environ==0.4.0
djangorestframework==3.3.2
gunicorn==19.4.5
psycopg2==2.6.1
six==1.10.0
whitenoise==2.0.6
